[Application] Move Data class to Data namespace.
