+++ src/Valkyrja/Application/Cli/Command/CacheCommand.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Valkyrja\Application\Cli\Command;
 
-use Valkyrja\Application\Data;
+use Valkyrja\Application\Data\Data;
 use Valkyrja\Application\Env;
 use Valkyrja\Cli\Interaction\Enum\ExitCode;
 use Valkyrja\Cli\Interaction\Factory\Contract\OutputFactory;
