+++ src/Valkyrja/Application/Contract/Application.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace Valkyrja\Application\Contract;
 
-use Valkyrja\Application\Data;
 use Valkyrja\Application\Data\Config;
+use Valkyrja\Application\Data\Data;
 use Valkyrja\Application\Env;
 use Valkyrja\Application\Support\Component;
 use Valkyrja\Container\Manager\Contract\Container;
