+++ src/Valkyrja/Application/Valkyrja.php
@@ -16,6 +16,7 @@ namespace Valkyrja\Application;
 use Override;
 use Valkyrja\Application\Contract\Application;
 use Valkyrja\Application\Data\Config;
+use Valkyrja\Application\Data\Data;
 use Valkyrja\Application\Exception\RuntimeException;
 use Valkyrja\Application\Support\Component;
 use Valkyrja\Cli\Routing\Data as CliData;
