+++ tests/Unit/Application/ApplicationTest.php
@@ -18,8 +18,8 @@ use Valkyrja\Application\Cli\Command\CacheCommand;
 use Valkyrja\Application\Cli\Command\ClearCacheCommand;
 use Valkyrja\Application\Component;
 use Valkyrja\Application\Contract\Application;
-use Valkyrja\Application\Data;
 use Valkyrja\Application\Data\Config;
+use Valkyrja\Application\Data\Data;
 use Valkyrja\Application\Env;
 use Valkyrja\Application\Exception\RuntimeException;
 use Valkyrja\Application\Valkyrja;
