+++ tests/Unit/Application/Cli/Command/CacheCommandTest.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Valkyrja\Tests\Unit\Application\Cli\Command;
 
 use Valkyrja\Application\Cli\Command\CacheCommand;
-use Valkyrja\Application\Data;
+use Valkyrja\Application\Data\Data;
 use Valkyrja\Cli\Interaction\Factory\OutputFactory;
 use Valkyrja\Cli\Routing\Collection\Collection as CliCollection;
 use Valkyrja\Container\Manager\Container;
